Patching over changes from a localized Git copy of the project. Unfortunately, the majority of these changes were irrecoverable due to compatibility problems with Git, so here's a complete log of what went into this:
commit 46b2a61068071075d5dc520f6e27a5de3b63f59a
Author: Thomas Robinson <trobinso@koala.local>
Date:   Thu Aug 11 11:41:23 2011 -0700

    Intelligibility update: made several nested print statements easier to understand

commit 667d2906e6c3477a43380f0ba165ecd4f35597b1
Author: Thomas Robinson <trobinso@koala.local>
Date:   Wed Aug 10 17:12:48 2011 -0700

    Updating tests cases, commenting out the few that seg fault for now (because of reliance on getopt_long; hey, it's not _our_ fault it crashes!)

commit 5c42fef5bb059336f61089514f35f6b5e3564b0a
Author: Thomas Robinson <trobinso@koala.local>
Date:   Wed Aug 10 15:44:04 2011 -0700

    Stylistic changes to enforce OTBS

commit 6378df44b734f46b9534e2629bcaa8e278bae37b
Author: Thomas Robinson <trobinso@koala.local>
Date:   Wed Aug 10 15:39:27 2011 -0700

    Removed tab characters from all source and test files, since Emacs M-x indent-region doesn't do that for us. Also fixed a few print messages to maintain a consistent look for the tool.

commit 191579033dd0127524591472cadf145893a0fbc7
Merge: 6900fca 54ef92a
Author: Thomas Robinson <trobinso@koala.local>
Date:   Wed Aug 10 15:15:38 2011 -0700

    Merge ssh://zucchini/local/projects/rf-ace-read-only.svn
     (Author's note: yes this is a local machine. No, you don't have access.)

    Conflicts:
        src/GBT.cpp
        src/datadefs.cpp
        src/datadefs.hpp
        src/node.cpp
        src/randomforest.cpp
        src/rf_ace.cpp

commit 6900fca435c1df807a36af3031363a345301ecb9
Author: Thomas Robinson <trobinso@koala.local>
Date:   Wed Aug 10 12:41:28 2011 -0700

    And ditto.

commit 59f9bcb78d9225a2d661ed78c6f1be60b86c31b7
Author: Thomas Robinson <trobinso@koala.local>
Date:   Wed Aug 10 12:41:14 2011 -0700

    Ditto.

commit 339b720e3b847a3628a1ddddcd038348c7eded4e
Author: Thomas Robinson <trobinso@koala.local>
Date:   Wed Aug 10 12:40:11 2011 -0700

    Cleaning up the file index

commit 48bdb4f82d097cb2542a2f7e71843be7a36ad4b7
Author: Thomas Robinson <trobinso@koala.local>
Date:   Wed Aug 10 12:20:53 2011 -0700

    Source and comment cleanup, including a working rendition of argparse_test.hpp and other goodies. This marks the first stage of usable unit testing, of which many more tests shall follow.

    To issue tests on a UNIX-based system, simply issue "make test". For Windows machines, please consult the documentation here:
    http://www.cs.nmsu.edu/~jeffery/courses/371/cppunit/money_example.html

    Yes, I do expect people to read this and for this to be eventually documented. But just to make sure, here's a shebang tag for you crazy grep fans: !! DOCUMENTATION

commit c042ef49007d350aa91cfb5e4124786940f62d7f
Author: Thomas Robinson <trobinso@koala.local>
Date:   Mon Aug 8 17:01:28 2011 -0700

    Adding RF-ACE launcher for GenePattern, for singular runs on a local machine

commit 898b3bd897096d017307aa7dd9b7cca6a1fa459e
Author: Thomas Robinson <trobinso@koala.local>
Date:   Mon Aug 8 16:53:10 2011 -0700

    Updating datadefs.cpp to fix a bug WRT the inclusion of newline characters

commit 6190945610288320631b97b9bad5b8e451b15a39
Author: Thomas Robinson <trobinso@koala.local>
Date:   Fri Aug 5 17:01:23 2011 -0700

    (See previous commit message)

commit eedf9b1751529703cd1e1c1f1b9be2af1cd54f60
Author: Thomas Robinson <trobinso@koala.local>
Date:   Fri Aug 5 16:56:41 2011 -0700

    Updating #includes to correctly work on a standard GNU system

commit f73b5abbe1849270a9d3cee9aa6372bf4d989f3a
Author: Thomas Robinson <trobinso@koala.local>
Date:   Fri Aug 5 16:19:43 2011 -0700

    Adding additional documentation, fixing various memory corruption bugs associated with GNU getopt

commit c844f8077a3c3f52e7bb08fb74d3d2ee598a8e68
Author: Thomas Robinson <trobinso@koala.local>
Date:   Fri Aug 5 14:53:01 2011 -0700

    Fixed two minor bugs WRT allowed arguments and a missing break statement.

commit b86658e35128e503dc7ad91f2b8f0f980cf63901
Author: Thomas Robinson <trobinso@koala.local>
Date:   Fri Aug 5 14:46:28 2011 -0700

    Major structural edits to rf_ace.cpp, and a complete refactoring of ArgParse to depend upon GNU getopt_long instead of defining all of this logic by hand. This changes the assumptions of input parsing to a more standard form, which should be better for all involved.

commit dfda35bf49801703924c100513296f2d2192bba8
Author: Thomas Robinson <trobinso@koala.local>
Date:   Thu Aug 4 12:20:27 2011 -0700

    Testing Git SSH access through a trivial change to the repository

commit 5c645bf17693716fe72cd884b91afb77e5f43ecb
Author: Thomas Robinson <trobinso@koala.local>
Date:   Mon Aug 1 15:23:25 2011 -0700

    Updating datadefs with additional documentation and one added pair of assertions

commit 55a6524dcc7b9d95fd3470e64cca8b0bd50dd407
Author: Thomas Robinson <trobinso@koala.local>
Date:   Fri Jul 29 17:51:59 2011 -0700

    Pushing changes to datadefs, plus a few documentation additions to other files

commit 61ad4067353660ad3b3ec2d7849a03eb07104229
Author: Thomas Robinson <trobinso@koala.local>
Date:   Thu Jul 28 18:01:33 2011 -0700

    Committing additional shebang documentation pertinent to the wide-area refactor

commit 9aad1b5be843ea338086af4fe7caa0880bfbd48e
Author: Thomas Robinson <trobinso@koala.local>
Date:   Thu Jul 28 17:17:39 2011 -0700

    Including my shebang (!!) changes and a minor fix for a comment

commit 04b15325ebe4bacacea0c5db3ffae16070bd63b9
Author: Thomas Robinson <trobinso@koala.local>
Date:   Tue Jul 26 16:45:23 2011 -0700

    Cleaning up cases where my text reflow broke the control flow of statements by wedging them into comment lines.

commit 99f39d9ec5175ac22813014190f5a17c7940f50e
Author: Thomas Robinson <trobinso@koala.local>
Date:   Tue Jul 26 15:43:23 2011 -0700

    Reflowed text to conform to 1TBS with tabs at two spaces apiece. Some discrepancies exist, but this should cover most of the current codebase.

commit db2f1bb19a70c687de59b21a5a2b90783fa3c159
Author: Thomas Robinson <trobinso@koala.local>
Date:   Mon Jul 25 17:14:39 2011 -0700

    Adding Doxygen configuration to the tree

+++ trunk/src/GBT.hpp
@@ -45,15 +45,15 @@ private:
   void transformLogistic(vector<num_t>& prediction, vector<num_t>& probability);
 
   Treedata* treeData_;
-  size_t targetIdx_;      // The index of the true target in treeData_
+  size_t targetIdx_;        // The index of the true target in treeData_
   size_t nTrees_;
   size_t nMaxLeaves_;
   //size_t nMaxNodes_;
-  size_t nodeSize_;		// smallest node size
-  float shrinkage_;
-  float subSampleSize_;
+  size_t nodeSize_;         // smallest node size
+  double shrinkage_;
+  double subSampleSize_;
   size_t numClasses_;
-  size_t nLeavesCounter_; 	// counter for constructing each tree
+  size_t nLeavesCounter_;   // counter for constructing each tree
 
   //Stores the rootnodes for the forest. Access the child nodes by calling rootNodes_[treeIdx]->leftChild() etc.
   //rootNodes_[treeIdx]->hasChildren() returns true if the node has children
